MsmqMessageBuilder supports DeliverBy and MaxAttempts (when it is 1).
