+++ Rhino.ServiceBus.Tests/Rhino.ServiceBus.Tests.csproj
@@ -191,7 +191,7 @@
     </Compile>
     <Compile Include="BugWithLogTest.cs" />
     <Compile Include="BusSubscriptionTests.cs" />
-    <Compile Include="CanCustomizeHeadersWithMsmq.cs" />
+    <Compile Include="CanCustomizeOutgoingMessagesWithMsmq.cs" />
     <Compile Include="Containers\Autofac\Can_host_in_another_app_domain.cs" />
     <Compile Include="Containers\Autofac\ContainerTests.cs" />
     <Compile Include="Containers\Autofac\QueueCreationTests.cs" />
